Move common camera code to a library.
