+++ camera/camera-utils/src/main/cpp/include/ndksamples/camera/camera_utils.h
