+++ camera/camera-utils/src/main/cpp/include/ndksamples/camera/native_debug.h
