+++ settings.gradle
@@ -25,6 +25,7 @@ include(":audio-echo:app")
 include(":base")
 include(":bitmap-plasma:app")
 include(":camera:basic")
+include(":camera:camera-utils")
 include(":camera:texture-view")
 include(":endless-tunnel:app")
 include(":exceptions:app")
